add: seconds in dashboard clock

+++ database/migrations/2024_05_16_084155_create_devices_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
                     ->onUpdate('cascade')->onDelete('cascade');
             $table->string('name');
             $table->string('topic');
-            $table->integer('capacity');
+            $table->float('capacity');
             $table->timestamps();
         });
     }
